Find edges between nodes based on a degree of neighborhood

Finds edges (links) between nodes (sites) based on a degree of neighborhood. The nodes labels (argument nodes) will be ordered to find edges in a directional way (from upstream to downstream along a linear shape).

Note that the detection of edges is only based on the nodes labels (no explicit spatial detection). For instance, if nodes are labelled as S-01, S-02, ..., S-10, the most upstream site will be S-01 and the most downstream S-10.

With a degree of neighborhood of 1, a node will be linked to the first previous node and also to the first next node (undirected network). So, two edges will be detected. If directed = FALSE, this node will be only linked to first next node (directed network).

edges_list(nodes, degree = 1, self = FALSE, all = FALSE, directed = FALSE)

Arguments

nodes

a character vector of nodes (sites) labels.

degree

an integer of length 1. The number of neighbors used to define edges.

self

a logical of length 1. If TRUE, a node can be linked to itself. Default is FALSE.

all

a logical of length 1. If TRUE, the missing edges are also returned. Default is FALSE.

directed

a logical of length 1. If FALSE (default), symmetrical edges (e.g. S01-S02 and S02-S01) are returned. Otherwise only the first edge (e.g. S01-S02) is returned (according to direction of the nodes labels).

Value

A data.frame with four columns:

from: label of one of the two nodes of the edge

to: label of the other node of the edge

edge: 0 (no edge) or 1 (edge)

edge_id: label of the edge

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# List of edges with 1 degree of neighborhood ----
edges_list(adour_sites$"site")
#>    edge_id edge from   to
#> 1      E-1    1 S-01 S-02
#> 2      E-1    1 S-02 S-01
#> 3      E-2    1 S-02 S-03
#> 4      E-2    1 S-03 S-02
#> 5      E-3    1 S-03 S-04
#> 6      E-3    1 S-04 S-03
#> 7      E-4    1 S-04 S-05
#> 8      E-4    1 S-05 S-04
#> 9      E-5    1 S-05 S-06
#> 10     E-5    1 S-06 S-05
#> 11     E-6    1 S-06 S-07
#> 12     E-6    1 S-07 S-06
#> 13     E-7    1 S-07 S-08
#> 14     E-7    1 S-08 S-07
#> 15     E-8    1 S-08 S-09
#> 16     E-8    1 S-09 S-08
#> 17     E-9    1 S-09 S-10
#> 18     E-9    1 S-10 S-09