Convert an adjacency matrix to an edges list

Converts an adjacency matrix to an edges list. This function allows to create the same edges list as the one obtained with edges_list().

matrix_to_edges_list(x, all = FALSE)

Arguments

x

a matrix object. The adjacency matrix to be converted in edges list.

all

a logical value. If FALSE (default), removes missing edges.

Value

A data.frame with four columns:

edge_id: label of the edge

edge: 0 (no edge) or 1 (edge)

from: label of one of the two nodes of the edge

to: label of the other node of the edge

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Find edges with 1 degree of neighborhood (undirected network) ----
adour_edges <- edges_list(adour_sites$"site")
adour_edges
#>    edge_id edge from   to
#> 1      E-1    1 S-01 S-02
#> 2      E-1    1 S-02 S-01
#> 3      E-2    1 S-02 S-03
#> 4      E-2    1 S-03 S-02
#> 5      E-3    1 S-03 S-04
#> 6      E-3    1 S-04 S-03
#> 7      E-4    1 S-04 S-05
#> 8      E-4    1 S-05 S-04
#> 9      E-5    1 S-05 S-06
#> 10     E-5    1 S-06 S-05
#> 11     E-6    1 S-06 S-07
#> 12     E-6    1 S-07 S-06
#> 13     E-7    1 S-07 S-08
#> 14     E-7    1 S-08 S-07
#> 15     E-8    1 S-08 S-09
#> 16     E-8    1 S-09 S-08
#> 17     E-9    1 S-09 S-10
#> 18     E-9    1 S-10 S-09

# Get adjacency matrix ----
adour_adj_matrix <- adjacency_matrix(adour_edges)
adour_adj_matrix
#>      S-01 S-02 S-03 S-04 S-05 S-06 S-07 S-08 S-09 S-10
#> S-01    0    1    0    0    0    0    0    0    0    0
#> S-02    1    0    1    0    0    0    0    0    0    0
#> S-03    0    1    0    1    0    0    0    0    0    0
#> S-04    0    0    1    0    1    0    0    0    0    0
#> S-05    0    0    0    1    0    1    0    0    0    0
#> S-06    0    0    0    0    1    0    1    0    0    0
#> S-07    0    0    0    0    0    1    0    1    0    0
#> S-08    0    0    0    0    0    0    1    0    1    0
#> S-09    0    0    0    0    0    0    0    1    0    1
#> S-10    0    0    0    0    0    0    0    0    1    0

# Convert back to edges list ----
new_edges <- matrix_to_edges_list(adour_adj_matrix)
new_edges
#>    edge_id edge from   to
#> 1      E-1    1 S-01 S-02
#> 2      E-1    1 S-02 S-01
#> 3      E-2    1 S-02 S-03
#> 4      E-2    1 S-03 S-02
#> 5      E-3    1 S-03 S-04
#> 6      E-3    1 S-04 S-03
#> 7      E-4    1 S-04 S-05
#> 8      E-4    1 S-05 S-04
#> 9      E-5    1 S-05 S-06
#> 10     E-5    1 S-06 S-05
#> 11     E-6    1 S-06 S-07
#> 12     E-6    1 S-07 S-06
#> 13     E-7    1 S-07 S-08
#> 14     E-7    1 S-08 S-07
#> 15     E-8    1 S-08 S-09
#> 16     E-8    1 S-09 S-08
#> 17     E-9    1 S-09 S-10
#> 18     E-9    1 S-10 S-09

# Check ----
identical(adour_edges, new_edges)
#> [1] TRUE