Create a nodes-by-edges matrix

Creates a nodes-by-edges matrix that will be used by adespatial::aem(). This function creates the same output as adespatial::aem.build.binary() but works in a different way: it's only based on nodes labels (not on coordinates). Also, this function adds labels to nodes and edges.

nodes_by_edges_matrix(edges)

Arguments

edges

a data.frame with the following three columns: from (the first node of the edge), to (the second node of the edge), and edge (presence/absence of the edge).

Value

A list of two elements:

se.mat: the nodes-by-edges matrix of dimensions n x k, where n is the number of nodes and k the number of edges (including the edge between the fictitious origin and the first site);

edges: a data.frame of edges list.

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Retrieve nodes (from nodes vector) ----
adour_nodes <- nodes_list(adour_sites$"site")

# Find edges with 1 degree of neighborhood ----
adour_edges <- edges_list(adour_nodes, directed = TRUE)

# Create nodes-by-edges matrix ----
adour_matrix <- nodes_by_edges_matrix(adour_edges)
adour_matrix
#> $se.mat
#>      E-01 E-02 E-03 E-04 E-05 E-06 E-07 E-08 E-09 E-10
#> S-01    1    0    0    0    0    0    0    0    0    0
#> S-02    1    1    0    0    0    0    0    0    0    0
#> S-03    1    1    1    0    0    0    0    0    0    0
#> S-04    1    1    1    1    0    0    0    0    0    0
#> S-05    1    1    1    1    1    0    0    0    0    0
#> S-06    1    1    1    1    1    1    0    0    0    0
#> S-07    1    1    1    1    1    1    1    0    0    0
#> S-08    1    1    1    1    1    1    1    1    0    0
#> S-09    1    1    1    1    1    1    1    1    1    0
#> S-10    1    1    1    1    1    1    1    1    1    1
#> 
#> $edges
#>      from   to
#> E-01    0 S-01
#> E-02 S-01 S-02
#> E-03 S-02 S-03
#> E-04 S-03 S-04
#> E-05 S-04 S-05
#> E-06 S-05 S-06
#> E-07 S-06 S-07
#> E-08 S-07 S-08
#> E-09 S-08 S-09
#> E-10 S-09 S-10
#>