Create a spatial weights matrix

Creates a spatial weights matrix by multiplying an adjacency (connectivity) matrix (see adjacency_matrix()) and an edges weight matrix (see edges_weight_matrix()). Resulting spatial weights equal to 0 will be replaced by 4 x max(w), where max(w) is the maximal weight in the matrix.

spatial_weights_matrix(x, y)

Arguments

x

an adjacency matrix of dimensions n x n, where n is the number of nodes (sites). The output of adjacency_matrix().

y

an edges weight matrix of dimensions n x n, where n is the number of nodes (sites). The output of edges_weight_matrix().

Value

A spatial weights matrix of dimensions n x n, where n is the number of nodes (sites).

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "bridge")
adour_sites  <- read.csv(path_to_file)

# Convert sites to sf object (POINTS) ----
adour_sites_sf <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

# Retrieve nodes (from nodes vector) ----
adour_nodes <- nodes_list(adour_sites$"site")

# Find edges with 1 degree of neighborhood (undirected) ----
adour_edges <- edges_list(adour_nodes)

# Get adjacency matrix ----
adour_adjacency <- adjacency_matrix(adour_edges)
adour_adjacency
#>      S-01 S-02 S-03 S-04 S-05 S-06 S-07 S-08 S-09 S-10
#> S-01    0    1    0    0    0    0    0    0    0    0
#> S-02    1    0    1    0    0    0    0    0    0    0
#> S-03    0    1    0    1    0    0    0    0    0    0
#> S-04    0    0    1    0    1    0    0    0    0    0
#> S-05    0    0    0    1    0    1    0    0    0    0
#> S-06    0    0    0    0    1    0    1    0    0    0
#> S-07    0    0    0    0    0    1    0    1    0    0
#> S-08    0    0    0    0    0    0    1    0    1    0
#> S-09    0    0    0    0    0    0    0    1    0    1
#> S-10    0    0    0    0    0    0    0    0    1    0

# Compute distances between pairs of sites along the Adour river ----
adour_dists <- distance_euclidean(adour_sites_sf)

# Create Edges weight matrix ----
adour_weights <- edges_weight_matrix(adour_dists)
adour_weights
#>           S-01      S-02       S-03       S-04     S-05     S-06     S-07
#> S-01      0.00  12278.29  23919.630  33207.477 47774.67 65640.77 76743.68
#> S-02  12278.29      0.00  11735.397  21298.857 35802.53 53533.55 64465.47
#> S-03  23919.63  11735.40      0.000   9786.038 24125.69 41798.64 52963.81
#> S-04  33207.48  21298.86   9786.038      0.000 14567.62 32501.89 44678.83
#> S-05  47774.67  35802.53  24125.688  14567.619     0.00 18023.36 31308.31
#> S-06  65640.77  53533.55  41798.645  32501.885 18023.36     0.00 15572.54
#> S-07  76743.68  64465.47  52963.814  44678.833 31308.31 15572.54     0.00
#> S-08  90280.92  78043.77  66789.510  58974.417 46018.95 30181.27 14834.32
#> S-09 101265.81  89278.65  78653.676  71864.676 60142.11 45826.85 30259.12
#> S-10 124742.69 114045.22 105473.228 101213.038 92665.19 82008.65 67179.49
#>          S-08      S-09      S-10
#> S-01 90280.92 101265.81 124742.69
#> S-02 78043.77  89278.65 114045.22
#> S-03 66789.51  78653.68 105473.23
#> S-04 58974.42  71864.68 101213.04
#> S-05 46018.95  60142.11  92665.19
#> S-06 30181.27  45826.85  82008.65
#> S-07 14834.32  30259.12  67179.49
#> S-08     0.00  16477.81  55447.54
#> S-09 16477.81      0.00  39509.30
#> S-10 55447.54  39509.30      0.00

# Create Spatial weight matrix ----
spatial_weights_matrix(adour_adjacency, adour_weights)
#>           S-01      S-02       S-03       S-04      S-05      S-06      S-07
#> S-01 158037.20  12278.29 158037.200 158037.200 158037.20 158037.20 158037.20
#> S-02  12278.29 158037.20  11735.397 158037.200 158037.20 158037.20 158037.20
#> S-03 158037.20  11735.40 158037.200   9786.038 158037.20 158037.20 158037.20
#> S-04 158037.20 158037.20   9786.038 158037.200  14567.62 158037.20 158037.20
#> S-05 158037.20 158037.20 158037.200  14567.619 158037.20  18023.36 158037.20
#> S-06 158037.20 158037.20 158037.200 158037.200  18023.36 158037.20  15572.54
#> S-07 158037.20 158037.20 158037.200 158037.200 158037.20  15572.54 158037.20
#> S-08 158037.20 158037.20 158037.200 158037.200 158037.20 158037.20  14834.32
#> S-09 158037.20 158037.20 158037.200 158037.200 158037.20 158037.20 158037.20
#> S-10 158037.20 158037.20 158037.200 158037.200 158037.20 158037.20 158037.20
#>           S-08      S-09     S-10
#> S-01 158037.20 158037.20 158037.2
#> S-02 158037.20 158037.20 158037.2
#> S-03 158037.20 158037.20 158037.2
#> S-04 158037.20 158037.20 158037.2
#> S-05 158037.20 158037.20 158037.2
#> S-06 158037.20 158037.20 158037.2
#> S-07  14834.32 158037.20 158037.2
#> S-08 158037.20  16477.81 158037.2
#> S-09  16477.81 158037.20  39509.3
#> S-10 158037.20  39509.30 158037.2